+++ quantum_debugger/algorithms/__init__.py
@@ -50,6 +50,7 @@ from .randomized_benchmarking import (
 )
 from .arithmetic import qft_add, quantum_adder
 from .maxcut import solve_maxcut, brute_force_maxcut
+from .state_preparation import ghz_state, w_state, graph_state
 
 __all__ = [
     # QFT
@@ -111,4 +112,8 @@ __all__ = [
     # QAOA MaxCut solver
     "solve_maxcut",
     "brute_force_maxcut",
+    # State preparation
+    "ghz_state",
+    "w_state",
+    "graph_state",
 ]